# docs/technical/openapi-hcb-fastify-snapshot.yaml
...
openapi: 3.0.3
info:
  title: HCB Admin Portal — Fastify dev API
  version: "1.0.0"
  description: |
    Development surrogate; in-memory state seeded from src/data/*.json.
    Not all response schemas are fully specified here.
servers:
  - url: http://127.0.0.1:8091/api
    description: Default local (npm run server)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCatalogLifecycleStatus:
      type: string
      description: |
        Catalogue product lifecycle for Fastify dev API. Create defaults to approval_pending when omitted.
        Reject or request-changes on the linked approval sets draft.
      enum:
        - approval_pending
        - pending
        - active
        - draft
    ConsortiumLifecycleStatus:
      type: string
      description: |
        Consortium row status. Omitting status on POST defaults to approval_pending (enqueues approval).
        POST with status active skips the queue.
      enum:
        - approval_pending
        - pending
        - active
        - inactive
paths:
...
  /v1/products:
    get:
      security:
        - bearerAuth: []
      summary: Paged product catalogue
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer }
      responses:
        "200":
          description: Paged products
    post:
      security:
        - bearerAuth: []
      summary: Create catalogue product
      description: |
        Requires non-empty `name`. Optional body fields include `description`, `status`, `price`, `pricingModel`,
        `packetIds`, `packetConfigs`, `enquiryConfig`.
        If `status` is `approval_pending` (or `pending`), the server also prepends an approvals item with
        `type: product` and `metadata.productId` pointing at the new id, and records `PRODUCT_CREATE` in the audit log.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                status:
                  $ref: "#/components/schemas/ProductCatalogLifecycleStatus"
                packetIds:
                  type: array
                  items: { type: string }
                packetConfigs: { type: array, items: { type: object } }
                enquiryConfig: { type: object }
      responses:
        "200":
          description: Created product (summary fields + id)
        "400":
          description: Validation (e.g. missing name)
  /v1/products/packet-catalog:
    get:
      security:
        - bearerAuth: []
      summary: Data product packet catalogue (Schema Mapper source types)
      description: |
        Returns `productCatalogPacketOptions` from the dev seed (`data-products.json`): each packet has a
        Schema Mapper `sourceType`, `category` group, field list, **`derivedFields`** (Configure → Derived tab),
        and `previewKey`. The create/edit product form uses this payload (with client-side filtering) so labels
        align with the Schema Mapper agent. Per-packet **Configure** in the SPA loads **raw paths** from
        `GET /v1/schema-mapper/schemas/source-type-fields` (merged with packet-only `fields` from this catalogue)
        and **Sources** from `GET /v1/schema-mapper/schemas?sourceType=`.
      responses:
        "200":
          description: Packet option list
          content:
            application/json:
              schema:
                type: object
                properties:
                  options:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        label: { type: string }
                        description: { type: string }
                        category: { type: string }
                        sourceType: { type: string }
                        previewKey: { type: string }
                        fields:
                          type: array
                          items: { type: string }
                        derivedFields:
                          type: array
                          items: { type: string }
                          description: Derived field names for this catalogue packet (product form Configure modal)
  /v1/products/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Product detail
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product including packetIds/packetConfigs/enquiryConfig when stored
    patch:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated product
    delete:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /v1/consortiums:
    get:
      security:
        - bearerAuth: []
      summary: Paged consortium list
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: size
          schema: { type: integer }
      responses:
        "200":
          description: Paged consortiums
    post:
      security:
        - bearerAuth: []
      summary: Create consortium
      description: |
        Requires `name`. Default `status` is `approval_pending` when omitted, which prepends an approvals row
        with `type: consortium` and `metadata.consortiumId`. Send `status: active` to skip the queue.
        Optional `members` (`institutionId` only) and `dataPolicy` (`dataVisibility` only: `full` \| `masked_pii` \| `derived`) are persisted; `shareLoanData`, `shareRepaymentHistory`, and `allowAggregation` in the body are ignored. Members populate `state.consortiumMembers`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                status:
                  $ref: "#/components/schemas/ConsortiumLifecycleStatus"
                dataPolicy:
                  type: object
                  description: Only dataVisibility is stored; other keys ignored.
                  properties:
                    dataVisibility:
                      type: string
                      enum: [full, masked_pii, derived]
                members:
                  type: array
                  items:
                    type: object
                    properties:
                      institutionId: { type: integer }
      responses:
        "200":
          description: Created consortium row
        "400":
          description: Validation (e.g. missing name)
  /v1/consortiums/{id}:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Consortium detail
    patch:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated consortium
    delete:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /v1/consortiums/{id}/members:
    get:
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Member rows for consortium
...